+++ tests/test_fair_estimations.py
@@ -154,8 +154,9 @@ def test_c_equivalence():
         c_equivalence(array, dims, eta=0, weights=None, func_args=None, func=None)
         == array.values.mean()
     )  # most things cancel out with eta=0
-    with pytest.raises(ZeroDivisionError):
-        c_equivalence(array, dims, eta=1, weights=None, func_args=None, func=None)
+    assert c_equivalence(
+        array, dims, eta=1, weights=None, func_args=None, func=None
+    ) == (np.exp(np.log(array).values.mean()))
     assert c_equivalence(
         array, dims, eta=10, weights=None, func_args=None, func=None
     ) == (np.divide(array ** (1 - 10), 1 - 10).values.mean() * (1 - 10)) ** (
@@ -189,19 +190,15 @@ def run_model_outputs(conf):
         formula=conf["formula"],
         year_range=conf["year_range"],
         year_start_pred=conf["year_start_pre"],
-        year_end_pred=conf["year_end_pre"],
         quantiles=conf["quantiles"],
         global_c=conf["global_c"],
         type_estimation=conf["type_estimation"],
-        extrap_formula="",  # deprecated when this test was written
-        extrap_year="",  # same
     )
 
     return out
 
 
 def test_model_outputs():
-
     """
     input cases variation :
     formula :